Split FlashID in header and Code file
